# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Premium Car Dealership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all test drive bookings
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
                  total:
                    type: integer
                required: [bookings, total]
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a test drive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                notes:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── TESTIMONIALS ──────────────────────────────────────────────────────────
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: published
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/BlogPost"
                  total:
                    type: integer
                required: [posts, total]
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── TEAM ─────────────────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  admin:
                    $ref: "#/components/schemas/AdminUser"
                required: [token, admin]
        "401":
          description: Unauthorized
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin user
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        preferredDate:
          type: string
        preferredTime:
          type: string
        location:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, phone, preferredDate, preferredTime, status, createdAt]
    CreateBookingBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        preferredDate:
          type: string
        preferredTime:
          type: string
        location:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required: [name, phone, preferredDate, preferredTime]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        category:
          type: string
        coverImage:
          type: string
          nullable: true
        isPublished:
          type: boolean
        publishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, slug, title, content, category, isPublished, createdAt]
    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        category:
          type: string
        coverImage:
          type: string
          nullable: true
        isPublished:
          type: boolean
      required: [title, content, category, isPublished]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
          nullable: true
      required: [id, username]
...