# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Premium Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  carId?: number | null;
  carTitle?: string | null;
  preferredDate: string;
  preferredTime: string;
  location?: string | null;
  message?: string | null;
  status: string;
  notes?: string | null;
  createdAt: string;
}